jarvis-agent/pyproject.toml
[build-system]
# Usamos setuptools para que el modo editable (`pip install -e`) cree bien el comando `jarvis`
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-agent"
version = "0.1.0"
description = "Jarvis-like CLI+Voice agent with tools (macOS) and GPT-5 Codex backend"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

# -------------------------
# Dependencias principales
# -------------------------
dependencies = [
  # --- Config / settings ---
  "python-dotenv>=1.0.1",        # Cargar .env
  "pydantic>=2.7.0",             # Settings tipados
  "pydantic-settings>=2.3.0",    # Settings desde env
  "platformdirs>=4.2.0",         # Rutas estándar (logs/cache) si lo necesitas

  # --- CLI / logs ---
  "rich>=13.7.1",                # UI de terminal (paneles, colores, progreso)

  # --- OpenAI (GPT-5 Codex) ---
  "openai>=1.40.0",              # SDK oficial OpenAI

  # --- Voz: captura de audio ---
  "sounddevice>=0.4.7",          # Micrófono cross-platform (Mac va muy bien)
  "numpy>=1.26.4",               # Necesario para audio buffers y utilidades

  # --- Wake word "Hey Jarvis" (Porcupine) ---
  # Porcupine funciona muy bien para activación por palabra clave.
  # pvrecorder es el capturador de mic recomendado para Porcupine.
  "pvporcupine>=3.0.4",
  "pvrecorder>=1.2.2",

  # --- Utilidades varias ---
  "requests>=2.32.3",            # HTTP básico (por si haces un buscador simple)
]

[project.optional-dependencies]
# Extras opcionales para herramientas futuras (no obligatorias ahora)
dev = [
  "pytest>=8.2.2",
  "ruff>=0.5.5",
  "mypy>=1.10.0",
]

# Si luego quieres automatizar navegador (para búsquedas y acciones en web)
browser = [
  "playwright>=1.46.0",
]

# Si luego haces embeddings/RAG local
rag = [
  "chromadb>=0.5.5",
]

[project.scripts]
# Esto crea el comando "jarvis" en tu entorno:
#   jarvis
# que ejecutará jarvis.main:main
jarvis = "jarvis.main:main"

[tool.setuptools]
# Estructura tipo src/ (tu paquete vive en src/jarvis)
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]